Add Bioconductor README badges

These functions add markdown text to the README to include all or individual badges from the Bioconductor landing page.

use_bioc_badges()

use_bioc_build_badge(which = c("release", "devel"))

use_bioc_support_badge()

use_bioc_rank_badge()

use_bioc_history_badge()

use_bioc_last_commit_badge()

use_bioc_dependencies_badge()

Arguments

which

Which branch to report the build status for. Possible values: "release" (default) or "devel".

Details

use_bioc_badges(): adds all badges listed below, mimicking the Bioconductor landing page

use_bioc_build_badge(): badge indicating Bioconductor build status for either the devel or release branch

use_bioc_support_badge(): Bioc support site activity in the last 6 months (answered posts / total posts)

use_bioc_rank_badge(): ranking by number of downloads

use_bioc_history_badge(): how long since the package was first in a released Bioconductor version (or if it is in devel only)

use_bioc_last_commit_badge(): time since last commit

use_bioc_dependencies_badge(): number of recursive dependencies needed to install the package

See also

Examples

if (FALSE) {
## Add all Bioc badges at once
use_bioc_badges()

## Add just the Bioc build status badge for devel
use_bioc_build_badge("devel")
}