Create a biocthis package template R scripts

This function creates the dev directory and in it, it creates R script files that you can follow to create your own Bioconductor-friendly R package. These R scripts include all the commands you need to get started =) The commands are designed to be run on the order given, though we also encourage you to check the latest utility functions in the usethis package.

use_bioc_pkg_templates(open = rlang::is_interactive())

Arguments

open

Open the newly created file for editing? Happens in RStudio, if applicable, or via utils::file.edit() otherwise. This argument has the same behavior as the open argument used in usethis functions.

Value

This function adds and/or replaces the dev files in your R package. That is:

dev/01_create_pkg.R

dev/02_git_github_setup.R

dev/03_core_files.R

dev/04_update.R

Details

For more details on how this function came to be, check: https://github.com/r-lib/actions/issues/84 https://github.com/r-lib/styler/issues/636 https://github.com/Bioconductor/BiocCheck/issues/57 https://github.com/Bioconductor/bioconductor.org/issues/54 https://github.com/r-lib/usethis/issues/1108 as well as the "biocthis developer notes" vignette https://lcolladotor.github.io/biocthis/articles/biocthis_dev_notes.html.

Examples

if (FALSE) {
## Run this function in your package
use_bioc_pkg_templates()
}

## Create an example package for illustrative purposes.
## Note: you do not need to run this for your own package!
pkgdir <- biocthis_example_pkg()
#> Package: biocthisexample
#> Title: What the Package Does (One Line, Title Case)
#> Version: 0.0.0.9000
#> Authors@R (parsed):
#>     * First Last <first.last@example.com> [aut, cre] (YOUR-ORCID-ID)
#> Description: What the package does (one paragraph).
#> License: `use_mit_license()`, `use_gpl3_license()` or friends to
#>     pick a license
#> Encoding: UTF-8
#> Roxygen: list(markdown = TRUE)
#> RoxygenNote: 7.2.3

## Create the biocthis templates
biocthis::use_bioc_pkg_templates()
#> ✔ Creating 'dev/'
#> ✔ Adding '^dev$' to '.Rbuildignore'
#> ✔ Writing 'dev/01_create_pkg.R'
#> ✔ Writing 'dev/02_git_github_setup.R'
#> ✔ Writing 'dev/03_core_files.R'
#> ✔ Writing 'dev/04_update.R'