Create a css with Bioconductor colors for pkgdown

This function creates the pkgdown/extra.css file with Bioconductor-style colors that will make your pkgdown documentation websites much cooler ^_^.

use_bioc_pkgdown_css()

Value

This function adds and/or replaces the pkgdown/extra.css file in your R package.

Details

For more details on how this function came to be, check https://github.com/lcolladotor/biocthis/issues/34.

Examples

if (FALSE) {
## Run this function in your package
use_bioc_support()
}

## Create an example package for illustrative purposes.
## Note: you do not need to run this for your own package!
pkgdir <- biocthis_example_pkg()
#> ✔ Setting active project to '/tmp/RtmpMfbZvR/biocthisexample'
#> Package: biocthisexample
#> Title: What the Package Does (One Line, Title Case)
#> Version: 0.0.0.9000
#> Authors@R (parsed):
#>     * First Last <first.last@example.com> [aut, cre] (YOUR-ORCID-ID)
#> Description: What the package does (one paragraph).
#> License: `use_mit_license()`, `use_gpl3_license()` or friends to
#>     pick a license
#> Encoding: UTF-8
#> Roxygen: list(markdown = TRUE)
#> RoxygenNote: 7.2.3
#> ✔ Setting active project to '/tmp/RtmpMfbZvR/biocthisexample'

## Create a css file to style your pkgdown website in Bioconductor colors
biocthis::use_bioc_pkgdown_css()
#> ✔ Creating 'pkgdown/'
#> ✔ Writing 'pkgdown/extra.css'