Create a biocthis-style DESCRIPTION template

This function is very similar to usethis::use_description() except that it uses a template from biocthis. This template includes some of the information expected in the DESCRIPTION file of a Bioconductor package. You can compare the output of this function to the one from usethis::use_description() to select the parts you want from each.

use_bioc_description(biocViews = "Software", report_bioc = TRUE)

Arguments

biocViews
A character() with the Bioconductor biocViews terms you want to use for your package. See https://bioconductor.org/packages/release/BiocViews.html for details. Note that the terms you choose have to be part of one of the main four trees: software, annotation, experiment or workflow.

report_bioc
A logical(1) indicating whether to set the BugReports to the Bioconductor Support website https://support.bioconductor.org/ or to the GitHub package issues page.

Value

This function adds and/or replaces the DESCRIPTION file in your R package.

Details

For details about the DESCRIPTION file guidelines for Bioconductor packages check http://bioconductor.org/developers/package-guidelines/#description.

Examples

if (FALSE) {
## Run this function in your package
use_bioc_description()
}

## Create an example package for illustrative purposes.
## Note: you do not need to run this for your own package!
pkgdir <- biocthis_example_pkg(use_git = TRUE)

#> ✔ Leaving 'DESCRIPTION' unchanged

#> Package: biocthisexample
#> Title: What the Package Does (One Line, Title Case)
#> Version: 0.0.0.9000
#> Authors@R (parsed):
#>     * First Last <first.last@example.com> [aut, cre] (YOUR-ORCID-ID)
#> Description: What the package does (one paragraph).
#> License: `use_mit_license()`, `use_gpl3_license()` or friends to
#>     pick a license
#> Encoding: UTF-8
#> LazyData: true
#> Roxygen: list(markdown = TRUE)
#> RoxygenNote: 7.1.1

#> ✔ Initialising Git repo

#> ✔ Adding '.Rproj.user', '.Rhistory', '.Rdata', '.httr-oauth', '.DS_Store' to '.gitignore'

## Create a template DESCRIPTION file that is Bioconductor-friendly
biocthis::use_bioc_description()

#> ✔ Leaving 'DESCRIPTION' unchanged

#> Package: biocthisexample
#> Title: What the Package Does (One Line, Title Case)
#> Version: 0.99.0
#> Date: 2020-12-17
#> Authors@R (parsed):
#>     * First Last <first.last@example.com> [aut, cre] (YOUR-ORCID-ID)
#> Description: What the package does (one paragraph).
#> License: Artistic-2.0
#> URL:
#> BugReports: https://support.bioconductor.org/t/biocthisexample
#> biocViews: Software
#> Encoding: UTF-8
#> LazyData: true
#> Roxygen: list(markdown = TRUE)
#> RoxygenNote: 7.1.1